Add "board" live view, with responsiveness for `key_down`
I have intentionally left some of the functionality out of this commit
so that Ryan Hinkle can try their hand at implementing some of the
remaining pieces.

The state of the code now closely mirrors what we had at the end of our
team session together during the November offsite.

Co-authored-by: Ryan Hinkle <ryan.hinkle@integral.io>